refactor(integrations): move github webhook flow out of communication
