+++ service/communication/urls.py
@@ -1,7 +1,1 @@
-from django.urls import path
-
-from communication.views import GithubWebhookView
-
-urlpatterns = [
-    path("webhook/github/", GithubWebhookView.as_view(), name="github-webhook"),
-]
+urlpatterns: list = []
