+++ service/integrations/admin.py
@@ -1,0 +1,1 @@
+"""Admin registrations for the integrations app."""
