+++ service/integrations/helpers.py
