+++ service/integrations/models.py
@@ -1,0 +1,1 @@
+"""Models for the integrations app."""
