+++ service/integrations/tests.py
@@ -1,0 +1,1 @@
+"""Tests for the integrations app."""
